requirements.txt
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
plotly==5.17.0
python-dotenv==1.0.0
langchain==0.1.4
langchain-experimental==0.0.47
langchain-openai==0.0.1
openai==1.3.5
hashlib
datetime